deepseek/requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
pymongo
pyyaml
requests